+++ Runtime/Cryptographers/CryptographerFactory.cs
@@ -1,0 +1,15 @@
+namespace ActionCode.Persistence
+{
+    public static class CryptographerFactory
+    {
+        public static ICryptographer Create(CryptographerType type, string key)
+        {
+            return type switch
+            {
+                CryptographerType.None => null,
+                CryptographerType.AES => new AESCryptographer(key),
+                _ => null,
+            };
+        }
+    }
+}