+++ Runtime/Cryptographers/CryptographerType.cs
@@ -1,0 +1,8 @@
+namespace ActionCode.Persistence
+{
+    public enum CryptographerType
+    {
+        None = -1,
+        AES
+    }
+}