+++ Runtime/Serializers/SerializerType.cs
@@ -1,0 +1,10 @@
+namespace ActionCode.Persistence
+{
+    public enum SerializerType
+    {
+        JsonUtility,
+        JsonNewtonsoft,
+        Xml,
+        Binary
+    }
+}